\docType{data}
\name{cellCycleKO}
\alias{cellCycleKO}
\title{A dataset based on Reimand \emph{et al} and Hu \emph{et al}. It contains lists
yeast genes that were most influenced by 12 cell cycle related transcription factor
knockouts.
The dataset is a list with 3 slots
\enumerate{
	\item \code{gl} - set of gene lists in a format suitable for
\code{\link{aggregateRanks}};
	\item \code{N} - number of yeast genes;
	\item \code{ref} - reference list of cell cycle related genes taken from  de
Lichtenberg \emph{et al}.
}}
\description{
  A dataset based on Reimand \emph{et al} and Hu \emph{et
  al}. It contains lists yeast genes that were most
  influenced by 12 cell cycle related transcription factor
  knockouts. The dataset is a list with 3 slots \enumerate{
  \item \code{gl} - set of gene lists in a format suitable
  for \code{\link{aggregateRanks}}; \item \code{N} - number
  of yeast genes; \item \code{ref} - reference list of cell
  cycle related genes taken from de Lichtenberg \emph{et
  al}. }
}
\author{
  Raivo Kolde <rkolde@gmail.com>
}
\references{
  Reimand, J., Vaquerizas, J. M., Todd, A. E., Vilo, J.,
  and Luscombe, N. M. (2010). "Comprehensive reanalysis of
  transcription factor knockout expression data in
  saccharomyces cerevisiae reveals many new targets.
  Nucleic Acids Res."

  Hu, Z., Killion, P. J., and Iyer, V. R. (2007). "Genetic
  reconstruction of a functional transcriptional regulatory
  network." Nat. Genet., 39(5), 683-7

  de Lichtenberg, U., Jensen, L. J., Fausboll, A., Jensen,
  T. S., Bork, P., and Brunak, S. (2005). "Comparison of
  computational methods for the identification of cell
  cycle- regulated genes. Bioinformatics, 21(7), 1164-71."
}
\keyword{data}
\keyword{datasets}

